Chronos partituta updated
git-svn-id: http://b035a0a.online-server.cloud/svn/ITDevelopment@14166 21205d9f-8918-2240-a3d5-c39064a2c852

+++ Include/imtsdlgenv2/CSdlClassGqlModificatorComp.cpp
@@ -267,7 +267,7 @@ bool CSdlClassGqlModificatorComp::AddContainerValueCheckConditionBegin(QTextStre
 			stream << QStringLiteral("QString");
 		}
 		else if (convertedType == QStringLiteral("int")){
-			stream << QStringLiteral("Double");
+			stream << QStringLiteral("Int");
 		}
 		else if (convertedType == QStringLiteral("long")){
 			stream << QStringLiteral("Long");
